requirements.txt
django==5.2.6
djangorestframework==3.16.1
django-rest-swagger==2.2.0
drf-yasg==1.21.10
gunicorn==23.0.0
python-decouple==3.8
psycopg2-binary==2.9.10
django-cors-headers==4.8.0
celery==5.5.3
redis==6.4.0
django-celery-beat==2.8.1
flower==2.0.1
django-redis==6.0.0
pylint==3.3.8
pylint-django==2.6.1
dnspython==2.8.0
tldextract==5.3.0
